1. 在小米9上测试发现最佳比例和单独benchmark得到的比例不同 2. 手动调整小米9上的比例为：0.75 0.7 0.7

+++ source/backend/codl/CoDLBackend.cpp
@@ -38,7 +38,7 @@ CoDLPartitionStrategy::CoDLPartitionStrategy(const std::string &jsonFile) {
     int n = obj["n"].GetInt();
     int m = obj["m"].GetInt();
     int k = obj["k"].GetInt();
-    float ratio = obj["ratio"].GetInt() * 0.1f;
+    float ratio = obj["ratio"].GetFloat() * 0.1f;
     auto dim = static_cast<CoDLNodePartitionParam::PartDim>(obj["dim"].GetInt());
 
     CoDLNodePartitionParam param = {dim, ratio};
